♻️ www_entry_point decorator -> www_som.helpers
